Add into_key() to Entry (was only on VacantEntry)
Before this change one needed to clone() key before uses of entry
api that might return OccupiedEntry. Now you can do .into_key() and
use the key again for something else.

The enabler for this was that OccupiedEntry's key was an Option<K>,
but was always used as Some(). Fixing this to be K made into_key()
pretty simple to add.

Tested with cargo test